decouple TypeAnalysis and code refactor
