+++ rap/src/analysis/core.rs
@@ -1,3 +1,4 @@
 pub mod alias;
 pub mod value_flow;
 pub mod control_flow;
+pub mod heap_item;
